Move dependent generate to assessRelevance
